Fix folder renaming issues
