+++ application/search/tests/ccs_comparison_search/test_service.py
@@ -1,4 +1,4 @@
-from ...ccs.service import Service
+from ...ccs_comparison_search.service import Service
 
 
 class TestService:
