+++ application/search/tests/test_search.py
@@ -1,6 +1,6 @@
 from json import dumps
 from unittest.mock import call, MagicMock, patch
-
+from .conftest import SEARCH_REQUEST
 from aws_lambda_powertools.utilities.data_classes import APIGatewayProxyEventV2
 from aws_lambda_powertools.utilities.typing.lambda_context import LambdaContext
 
@@ -20,7 +20,15 @@ def test_lambda_handler(
     response = lambda_handler(search_request, lambda_context)
     # Assert
     assert response["statusCode"] == 200
-    assert response["body"] == dumps({"search_one": return_value, "search_two": return_value})
+
+    assert response["body"] == dumps(
+        {
+            "search_one": return_value,
+            "search_one_environment": SEARCH_REQUEST["search_one"]["search_environment"],
+            "search_two": return_value,
+            "search_two_environment": SEARCH_REQUEST["search_two"]["search_environment"],
+        }
+    )
     mock_check_capacity_summary_search.assert_has_calls(
         [
             call(
@@ -42,7 +50,7 @@ def test_lambda_handler(
                 symptom_discriminator_list=[1, 2, 3],
                 search_distance=1,
                 gender="male",
-                search_environment="test",
+                search_environment="test2",
             ),
             call().search(),
         ]
